Conflicts:
	org.caleydo.core/src/org/caleydo/core/data/collection/table/Table.java
	org.caleydo.internal/launch/Caleydo Mac Experimental.launch
	org.caleydo.view.stratomex/src/org/caleydo/view/stratomex/tourguide/internal/TemplateHighlightRenderer.java
	org.caleydo.view.tourguide/src/main/java/org/caleydo/view/tourguide/impl/algorithm/LogRank.java
	org.caleydo.view.tourguide/src/main/java/org/caleydo/view/tourguide/impl/algorithm/Statistics.java
	org.caleydo.view.tourguide/src/main/java/org/caleydo/view/tourguide/internal/model/MaxGroupCombiner.java
	org.caleydo.vis.lineup/src/main/java/org/caleydo/vis/lineup/model/mapping/JavaScriptFunctions.java

+++ org.caleydo.view.tourguide/src/main/java/org/caleydo/view/tourguide/impl/algorithm/LogRank.java
@@ -87,7 +87,7 @@ public class LogRank implements IGroupAlgorithm {
 		double r = Statistics.logRank(as, asurvived, bs, bsurvived);
 		if (Double.isInfinite(r))
 			return Double.NaN;
-		return r;
+		return (float) r;
 	}
 
 	private Pair<List<Double>, Integer> getValues(Iterable<Integer> a, Integer col) {
